pyproject.toml
[project]
name = "python_utils"
description = "Collection of utilities for Python projects."
version = "1.0.0"
authors = [{ name = "Mathieu Gueribout", email = "gueribout.mathieu@protonmail.com" }]
maintainers = [{ name = "Mathieu Gueribout", email = "gueribout.mathieu@protonmail.com" }]
readme = "README.md"
requires-python = ">=3.11, <3.12"
keywords = ["python", "utils", "tools"]
classifiers = ["Programming Language :: Python :: 3"]

dependencies = [
    "alembic==1.14.*",
    "coverage==7.6.*",
    "docker==7.1.*",
    "fastapi==0.115.*",
    "httpx==0.28.*",
    "pre-commit==4.0.*",
    "psycopg2-binary==2.9.*",
    "pyright==1.1.*",
    "pytest==8.3.*",
    "pytest-mock==3.14.*",
    "python-dotenv==1.0.*",
    "python-json-logger==3.2.*",
    "requests==2.32.*",
    "ruff==0.9.*",
    "sqlalchemy==2.0.*",
    "types-sqlalchemy-utils==1.1.*",
    "uuid6==2024.7.*",
    "uvicorn==0.34.*",
]

[project.urls]
repository = "https://github.com/gueriboutmathieu/python_utils"

[tool.uv]
package = true  # Install current package when running `uv sync`

[tool.coverage.run]
source = ["python_utils"]
relative_files = true

[tool.ruff]
target-version = "py311"
exclude = [
    ".git",
    ".github",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
fixable = ["ALL"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pyright]
typeCheckingMode = "strict"
include = ["python_utils"]
exclude = ["*.venv/**", ".github/*"]
